requirements.txt
pandas
openpyxl
numpy
opencv-python
tqdm
tensorflow
Pillow >= 8.0
matplotlib
tensorflow-datasets
sacred